Add routes using BrowserRouter, Routes, and Route; restructure folders
