+++ src/App.css
@@ -5,8 +5,9 @@
 	text-align: center;
 }
 
+/* 
 .app {
 	display: flex;
 	justify-content: center;
 	align-items: center;
-}
+} */
