+++ src/App.jsx
@@ -1,14 +1,20 @@
 /* eslint-disable no-unused-vars */
 import React from "react";
-import Form from "./components/form";
+import Home from "./components/Home/Home";
+import Form from "./components/Form/Form";
 import "./App.css";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 const App = () => {
 	return (
 		<>
-			<div className="app">
-				<Form />
-			</div>
+			<BrowserRouter>
+				<Routes>
+					<Route path="/" element={<Home />} />
+					<Route path="*" element={<Home />} />
+					<Route path="/contact-us" element={<Form />} />
+				</Routes>
+			</BrowserRouter>
 		</>
 	);
 };
