+++ src/components/Home/Home.jsx
@@ -1,0 +1,8 @@
+/* eslint-disable no-unused-vars */
+import React from "react";
+
+const Home = () => {
+	return <div>Home</div>;
+};
+
+export default Home;
